Build and warning fix

+++ Lumos/Source/Lumos/Platform/Vulkan/VKDescriptorSet.cpp
@@ -642,7 +642,7 @@ namespace Lumos
             LWARN("Uniform not found %d", int(binding));
         }
 
-        void VKDescriptorSet::SetUniformBufferData(u8 binding, void* data, float size)
+        void VKDescriptorSet::SetUniformBufferData(u8 binding, void* data, uint32_t size)
         {
             LUMOS_PROFILE_FUNCTION();
